RavenDB-19149 Fixing condition in MergerScheduler | Moving from IndexHadler into AdminIndexHandler
